requirements.txt
dj-database-url==0.4.0
Django==1.9.4
django-filebrowser==3.7.1
django-grappelli==2.8.1
django-haystack==2.4.1
django-leaflet==0.18.0
django-modeltranslation==0.11
django-registration==2.0.4
django-tinymce==2.3.0
django-widget-tweaks==1.4.1
easy-thumbnails==2.3
gunicorn==19.4.5
Pillow==3.2.0
psycopg2==2.6.1
pytz==2016.3
reportlab==3.3.0
requests==2.9.1
sqlparse==0.1.19
stripe==1.31.0
whitenoise==3.0
